@@ -162,8 +162,8 @@
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
         <configuration>
-          <source>9</source>
-          <target>9</target>
+          <source>10</source>
+          <target>10</target>
         </configuration>
       </plugin>
 
