@@ -14,11 +14,13 @@ import com.spring.boot.like.domain.Like;
 import com.spring.boot.like.domain.LikeRepository;
 import com.spring.boot.member.domain.Member;
 import com.spring.boot.post.domain.Post;
+import com.spring.boot.post.infrastructure.PostRepository;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import org.hibernate.NonUniqueResultException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -30,6 +32,8 @@ class LikeServiceTest extends IntegrationTest {
   private LikeService likeService;
   @Autowired
   private LikeRepository likeRepository;
+  @Autowired
+  private PostRepository postRepository;
 
   @Nested
   @DisplayName("좋아요가 이미 눌러져 있을 때")
@@ -63,6 +67,9 @@ class LikeServiceTest extends IntegrationTest {
       // then
       List<LikeDto> likes = likeService.getLikes(post.getId());
       assertThat(likes.size(), is(0));
+
+      Post findPost = postRepository.findById(post.getId()).orElseThrow();
+      assertThat(findPost.getPostLikeCount(), is(0));
     }
 
     @Test
@@ -91,6 +98,9 @@ class LikeServiceTest extends IntegrationTest {
       latch.await();
       List<LikeDto> likes = likeService.getLikes(post.getId());
       assertThat(likes.size(), is(0));
+
+      Post findPost = postRepository.findById(post.getId()).orElseThrow();
+      assertThat(findPost.getPostLikeCount(), is(0));
     }
   }
 
@@ -100,12 +110,12 @@ class LikeServiceTest extends IntegrationTest {
 
     @Test
     @DisplayName("취소 요청하면 예외 발생한다.")
-    public void 취소_요청(){
+    public void 취소_요청() {
       Member member = saveMember("member@email.com");
       Member writer = saveMember("writer@email.com");
       Post post = savePost(writer);
 
-      assertThrows(NotFoundException.class, ()->{
+      assertThrows(NotFoundException.class, () -> {
         likeService.unlike(member.getId(), post.getId());
       });
     }
@@ -124,10 +134,13 @@ class LikeServiceTest extends IntegrationTest {
       Like findLike = like.get();
       assertThat(findLike.getMember().getId(), is(member.getId()));
       assertThat(findLike.getPost().getId(), is(post.getId()));
+
+      Post findPost = postRepository.findById(post.getId()).orElseThrow();
+      assertThat(findPost.getPostLikeCount(), is(1));
     }
 
     @Test
-    @DisplayName("동시에 여러번 좋아요 요청해도 좋아요 최종 개수 1개")
+    @DisplayName("동시에 좋아요 요청했을 때 좋아요 한 개만 저장된다.(중복 저장 x)")
     public void 동시에_여러개_요청() throws InterruptedException {
       // given
       Member member = saveMember("member@email.com");
@@ -136,12 +149,13 @@ class LikeServiceTest extends IntegrationTest {
 
       // when
       ExecutorService executorService = Executors.newFixedThreadPool(30);
-      CountDownLatch latch = new CountDownLatch(100);
-      for (int i = 0; i < 100; i++) {
+      int count = 30;
+      CountDownLatch latch = new CountDownLatch(count);
+      for (int i = 0; i < count; i++) {
         executorService.submit(() -> {
           try {
             likeService.like(member.getId(), post.getId());
-          }finally {
+          } finally {
             latch.countDown();
           }
         });
@@ -151,6 +165,9 @@ class LikeServiceTest extends IntegrationTest {
       latch.await();
       List<LikeDto> likes = likeService.getLikes(post.getId());
       assertThat(likes.size(), is(1));
+
+      Post findPost = postRepository.findById(post.getId()).orElseThrow();
+      assertThat(findPost.getPostLikeCount(), is(1));
     }
 
   }
